% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepdivedata-package.R
\docType{package}
\name{deepdivedata-package}
\alias{deepdivedata}
\alias{deepdivedata-package}
\title{deepdivedata: Data package for Deep Dive Into Data OI}
\description{
This package contains all datasets for coding tutorials associated with the Deep Dive into Data program created by the Centre for Ocean Ventures and Entrepreneurship (COVE).
}
\author{
\strong{Maintainer}: Emily Allen \email{emily.allen@coveocean.com}

}
\keyword{internal}
