% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempandtideData.R
\docType{data}
\name{harbourData}
\alias{harbourData}
\title{Temperature and tide data collected in the Halifax Harbour in May 2022 using RBR CTD sensor deployed on seabed platform Stella Maris
A CTD sensor was used to measured temperature and depth}
\format{
A tibble with 144 rows and 7 variables:
\describe{
\item{Timestamp}{Date and time of data collection}
\item{Temperature}{ocean temperature}
\item{Units}{ocean temperature measured in C}
\item{Depth}{depth}
\item{Units}{depth measured in m}
\item{Pressure}{pressure}
\item{Units}{pressure measured in dbar}
}
}
\source{
\url{https://github.com/deepdiveOI/Data}
}
\usage{
harbourData
}
\description{
Temperature and tide data collected in the Halifax Harbour in May 2022 using RBR CTD sensor deployed on seabed platform Stella Maris
A CTD sensor was used to measured temperature and depth
}
\keyword{datasets}
