% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasurface.R
\docType{data}
\name{seasurface}
\alias{seasurface}
\title{Data of global annual sea surface temperature anomalies from 1880-2020
Temperature anomalies are recorded in degrees Celsius, and the average sea surface temperature from 1971-2000 is used as the baseline.}
\format{
A tibble with 141 rows and 2 variables:
\describe{
\item{Year}{Year that temperatures were recorded}
\item{annual.anom}{The temperature anomaly in degrees Celcius based on the average annual sea surface temperature}
}
\source{
\url{https://www.statista.com/statistics/224893/land-and-ocean-temperature-anomalies-based-on-temperature-departure/#:~:text=In%202021%2C%20the%20land%20and,climate%20change%20than%20absolute%20temperature.}
}
}
\usage{
seasurface
}
\description{
Data of global annual sea surface temperature anomalies from 1880-2020
Temperature anomalies are recorded in degrees Celsius, and the average sea surface temperature from 1971-2000 is used as the baseline.
}
\keyword{datasets}
