% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormData.R
\docType{data}
\name{stormData}
\alias{stormData}
\title{Storm data collected by CTD sensors and the COVE weather station in November 2021
Various ocean and atmospheric variables give insight to how big storms affect the ocean}
\format{
A tibble with 144 rows and 7 variables:
\describe{
\item{Timestamp}{Date and time of data collection}
\item{Temperature}{ocean temperature}
\item{Units}{temperature measured in C}
\item{Depth}{depth}
\item{Units}{depth measured in m}
\item{Pressure}{pressure}
\item{Units}{pressure measured in dbar}
}
}
\source{
\url{https://github.com/deepdiveOI/Data}
}
\usage{
stormData
}
\description{
Storm data collected by CTD sensors and the COVE weather station in November 2021
Various ocean and atmospheric variables give insight to how big storms affect the ocean
}
\keyword{datasets}
