% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbidityData.R
\docType{data}
\name{turbidityData}
\alias{turbidityData}
\title{Turbidity data collected in the Halifax Harbour in 2022 using an Xeos sensor deployed on the seabed platform Stella Maris
Turbidity measurements are recorded in NTU, and the data were collected every hour from April 5-11, 2022}
\format{
A tibble with 169 rows and 3 variables:
\describe{
\item{Timestamp}{Date and hour that the data were recorded}
\item{Turbidity}{the turbidity measured at a given time}
\item{Units}{turbidity was recorded in NTU}
}
}
\source{
\url{https://github.com/deepdiveOI/Data}
}
\usage{
turbidityData
}
\description{
Turbidity data collected in the Halifax Harbour in 2022 using an Xeos sensor deployed on the seabed platform Stella Maris
Turbidity measurements are recorded in NTU, and the data were collected every hour from April 5-11, 2022
}
\keyword{datasets}
